Created two-stage buffering using data/process.
Slow machines were still seeing dropped samples so now we
use two buffers: a raw sample buffer filled during the
endpoint callback (which just saves them in an array),
and a processed sample buffer filled by the RawProcessor
callback. This buffer is flushed to disk when it fills,
and on exit.

Not sure of the sizes of the buffers though. Clearly the
data_fn callback needs to be reasonably sized to prevent
overflow on slow machines, but won't too big of a processed
sample buffer cause a big file write penalty periodically?